Revise Input Management
Input is now being handled by an input manager, triggering delegates that methods will listen to.

Each scene will have an initialization game object that tells the input manager what InputAction map it will have to control when the scene is first loaded. This is where methods in the scene will listen to and connect to the input manager events.

Event gets called, methods are invoked, and player can make input.