Adding Initialization Checks & Fixing Small Bug
Adding an initialization check for the initialization scene. Added a game manager object to handle the logic.

Realistically, this check shouldn't be needed, but it's better to be safe than sorry.

Each singleton initialized on the initialization scene will have a check to ensure that it is ready before the game starts.